add worker
